Program Jokian Bak sampah/requirements.txt
SpeechRecognition>=3.10.0
pyserial>=3.5
PyAudio>=0.2.14
